Refactored to have Vectors in my Sprite Class
TODOs:
 Flip method for Sprite
 All the AnimatedSpriteClass

+++ src/sample/Vector.java
@@ -37,10 +37,10 @@ class Vector {
         return v;
     }
 
-    public void setX(int _x){
+    public void setX(double _x){
         x = _x;
     }
-    public void setY(int _y){
+    public void setY(double _y){
         y = _y;
     }
